[NEW] account_advanced_protection_features: fusion of both modules account_move_protect_deleting and account_move_protect_reset_to_draft
