+++ account_advanced_protection_features/i18n/fr.po
@@ -1,19 +1,27 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# 	* account_move_protect_reset_to_draft
+# 	* account_advanced_protection_features
 #
 msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-10-14 14:53+0000\n"
-"PO-Revision-Date: 2025-06-06 15:47+0200\n"
+"POT-Creation-Date: 2024-03-26 13:13+0000\n"
+"PO-Revision-Date: 2024-03-26 13:13+0000\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
 "Plural-Forms: \n"
+
+#. module: account_move_protect_deleting
+#. odoo-python
+#: code:addons/account_move_protect_deleting/models/account_move.py:0
+#, python-format
+msgid "You cannot delete this account move because it has been posted."
+msgstr "Vous ne pouvez pas supprimer cette pièce comptable car elle à déjà été confirmée."
+
 
 #. module: account_move_protect_reset_to_draft
 #: model:ir.model,name:account_move_protect_reset_to_draft.model_account_invoice_send
