+++ account_advanced_protection_features/models/account_move.py
@@ -1,8 +1,18 @@
-from odoo import models, api, _
+from odoo import models, api, fields, _
 from odoo.exceptions import UserError
 
 class AccountMove(models.Model):
     _inherit = 'account.move'
+
+    sent_by_email = fields.Boolean()
+
+    def button_draft(self):
+        res = super(AccountMove, self).button_draft()
+        if self.sent_by_email:
+            raise UserError(_(
+                "You cannot reset to draft this invoice because it has been sent by email."
+            ))
+        return res
 
     @api.ondelete(at_uninstall=False)
     def _check_posted(self):
