--- account_move_protect_deleting/i18n/fr.po
@@ -1,23 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_move_protect_deleting
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-03-26 13:13+0000\n"
-"PO-Revision-Date: 2024-03-26 13:13+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_move_protect_deleting
-#. odoo-python
-#: code:addons/account_move_protect_deleting/models/account_move.py:0
-#, python-format
-msgid "You cannot delete this account move because it has been posted."
-msgstr "Vous ne pouvez pas supprimer cette pièce comptable car elle à déjà été confirmée."
