--- account_move_protect_reset_to_draft/README.rst
@@ -1,44 +1,0 @@
-===============
-account_move_protect_reset_to_draft
-===============
-
-Forbid to delete an invoice that has been already validated
-
-Installation
-============
-
-Use Odoo normal module installation procedure to install
-``account_move_protect_reset_to_draft``.
-
-Known issues / Roadmap
-======================
-
-None yet.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `our issues website <https://github.com/elabore-coop/account_move_protect_deleting/issues>`_. In case of
-trouble, please check there if your issue has already been
-reported. If you spotted it first, help us smashing it by providing a
-detailed and welcomed feedback.
-
-Credits
-=======
-
-Contributors
-------------
-
-* Clément Thomas
-
-Funders
--------
-
-The development of this module has been financially supported by:
-* Elabore (https://elabore.coop)
-
-
-Maintainer
-----------
-
-This module is maintained by Elabore.
