--- account_move_protect_reset_to_draft/__init__.py
@@ -1,2 +1,0 @@
-from . import models
-from . import wizard
