--- account_move_protect_reset_to_draft/__manifest__.py
@@ -1,34 +1,0 @@
-# Copyright 2024 Boris Gallet ()
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-{
-    "name": "account_move_protect_reset_to_draft",
-    "version": "16.0.1.0.0",
-    "author": "Elabore",
-    "website": "https://elabore.coop",
-    "maintainer": "Clément Thomas",
-    "license": "AGPL-3",
-    "category": "Tools",
-    "summary": "prohibits re-draft an invoice that has been sent by email",
-    # any module necessary for this one to work correctly
-    "depends": [
-        "base",
-        "account"
-    ],
-    "qweb": [],
-    "external_dependencies": {
-        "python": [],
-    },
-    # always loaded
-    "data": [
-    ],
-    # only loaded in demonstration mode
-    "demo": [],
-    "js": [],
-    "css": [],
-    "installable": True,
-    # Install this module automatically if all dependency have been previously
-    # and independently installed.  Used for synergetic or glue modules.
-    "auto_install": False,
-    "application": False,
-}
