--- account_move_protect_reset_to_draft/i18n/account_move_protect_reset_to_draft.pot
@@ -1,41 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_move_protect_reset_to_draft
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2024-10-14 14:52+0000\n"
-"PO-Revision-Date: 2024-10-14 14:52+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_move_protect_reset_to_draft
-#: model:ir.model,name:account_move_protect_reset_to_draft.model_account_invoice_send
-msgid "Account Invoice Send"
-msgstr ""
-
-#. module: account_move_protect_reset_to_draft
-#: model:ir.model,name:account_move_protect_reset_to_draft.model_account_move
-msgid "Journal Entry"
-msgstr ""
-
-#. module: account_move_protect_reset_to_draft
-#: model:ir.model.fields,field_description:account_move_protect_reset_to_draft.field_account_bank_statement_line__sent_by_email
-#: model:ir.model.fields,field_description:account_move_protect_reset_to_draft.field_account_move__sent_by_email
-#: model:ir.model.fields,field_description:account_move_protect_reset_to_draft.field_account_payment__sent_by_email
-msgid "Sent By Email"
-msgstr ""
-
-#. module: account_move_protect_reset_to_draft
-#. odoo-python
-#: code:addons/account_move_protect_reset_to_draft/models/account_move.py:0
-#, python-format
-msgid ""
-"You cannot reset to draft this invoice because it has been sent by email."
-msgstr ""
