--- account_move_protect_reset_to_draft/models/__init__.py
@@ -1,1 +1,0 @@
-from . import account_move
